Consolidated payment trigger tests, fixed commit issue.
