+++ .githooks/pre-commit
@@ -1,2 +1,2 @@
 #!/bin/sh
-npx lint-staged
+node ./node_modules/lint-staged/bin/lint-staged.js
