--- tests/integration-tests/triggerJsonValidation.test.ts
@@ -1,265 +1,0 @@
-import { prismaMock } from '../../prisma-local/mockedClient'
-import prisma from '../../prisma-local/clientInstance'
-import axios from 'axios'
-
-import { parseTriggerPostData } from '../../utils/validators'
-
-// Mock axios for external requests
-jest.mock('axios')
-const mockedAxios = axios as jest.Mocked<typeof axios>
-
-// Mock utils/validators to control JSON parsing behavior
-jest.mock('../../utils/validators', () => {
-  const originalModule = jest.requireActual('../../utils/validators')
-  return {
-    ...originalModule,
-    parseTriggerPostData: jest.fn()
-  }
-})
-const mockedParseTriggerPostData = parseTriggerPostData as jest.MockedFunction<typeof parseTriggerPostData>
-
-describe('Trigger JSON Validation Integration Tests', () => {
-  describe('Trigger Creation with JSON Validation', () => {
-    it('should reject trigger creation with invalid JSON during validation', async () => {
-      // This tests the existing validation during trigger creation
-      // which should catch most JSON issues before they reach execution
-
-      const invalidPostData = '{"amount": <amount>, "currency": <currency>' // Missing closing brace
-
-      // Mock the parsing to fail during creation validation
-      mockedParseTriggerPostData.mockImplementation(() => {
-        throw new SyntaxError('Unexpected end of JSON input')
-      })
-
-      expect(() => {
-        // This would be caught by parsePaybuttonTriggerPOSTRequest during creation
-        parseTriggerPostData({
-          userId: 'test-user',
-          postData: invalidPostData,
-          postDataParameters: {} as any
-        })
-      }).toThrow('Unexpected end of JSON input')
-
-      // Verify that the parsing was attempted
-      expect(mockedParseTriggerPostData).toHaveBeenCalled()
-    })
-
-    it('should allow trigger creation with valid JSON', async () => {
-      const validPostData = '{"amount": <amount>, "currency": <currency>}'
-      const expectedParsedData = { amount: 100, currency: 'XEC' }
-
-      // Mock successful parsing
-      mockedParseTriggerPostData.mockReturnValue(expectedParsedData)
-
-      const result = parseTriggerPostData({
-        userId: 'test-user',
-        postData: validPostData,
-        postDataParameters: {} as any
-      })
-
-      expect(result).toEqual(expectedParsedData)
-      expect(mockedParseTriggerPostData).toHaveBeenCalled()
-    })
-  })
-
-  describe('Trigger Execution Scenarios', () => {
-    beforeEach(() => {
-      jest.clearAllMocks()
-
-      // Setup common mocks
-      prismaMock.triggerLog.create.mockResolvedValue({
-        id: 1,
-        triggerId: 'test-trigger',
-        isError: false,
-        actionType: 'PostData',
-        data: '{}',
-        createdAt: new Date(),
-        updatedAt: new Date()
-      })
-      prisma.triggerLog.create = prismaMock.triggerLog.create
-    })
-
-    it('should demonstrate JSON validation flow differences', async () => {
-      // Test Case 1: Valid JSON - should proceed to network request
-      console.log('=== Test Case 1: Valid JSON ===')
-
-      mockedParseTriggerPostData.mockReturnValue({ amount: 100, currency: 'XEC' })
-      mockedAxios.post.mockResolvedValue({ data: { success: true } })
-
-      // Simulate the parsing (this would happen in postDataForTrigger)
-      try {
-        const result = parseTriggerPostData({
-          userId: 'user-123',
-          postData: '{"amount": <amount>, "currency": <currency>}',
-          postDataParameters: {} as any
-        })
-        console.log('✅ JSON parsing succeeded:', result)
-        console.log('✅ Network request would be made')
-      } catch (error) {
-        console.log('❌ Unexpected error:', error)
-      }
-
-      // Test Case 2: Invalid JSON - should fail early
-      console.log('\n=== Test Case 2: Invalid JSON ===')
-
-      mockedParseTriggerPostData.mockImplementation(() => {
-        throw new SyntaxError('Unexpected end of JSON input')
-      })
-
-      try {
-        parseTriggerPostData({
-          userId: 'user-123',
-          postData: '{"amount": <amount>, "currency": <currency>',
-          postDataParameters: {} as any
-        })
-        console.log('❌ Should not reach here')
-      } catch (error) {
-        console.log('✅ JSON parsing failed as expected:', (error as Error).message)
-        console.log('✅ Network request would NOT be made')
-      }
-
-      // Verify the behavior difference
-      expect(mockedParseTriggerPostData).toHaveBeenCalledTimes(2)
-    })
-
-    it('should log JSON validation errors with proper details', async () => {
-      const testCases = [
-        {
-          name: 'Missing closing brace',
-          postData: '{"amount": <amount>, "currency": <currency>',
-          expectedError: 'Unexpected end of JSON input'
-        },
-        {
-          name: 'Invalid property syntax',
-          postData: '{amount: <amount>, "currency": <currency>}',
-          expectedError: 'Expected property name'
-        },
-        {
-          name: 'Extra comma',
-          postData: '{"amount": <amount>,, "currency": <currency>}',
-          expectedError: 'Unexpected token'
-        }
-      ]
-
-      testCases.forEach(({ name, postData, expectedError }) => {
-        console.log(`\n=== Testing: ${name} ===`)
-
-        // Mock parsing to fail with specific error
-        mockedParseTriggerPostData.mockImplementation(() => {
-          const error = new SyntaxError(expectedError)
-          error.name = 'SyntaxError'
-          throw error
-        })
-
-        try {
-          parseTriggerPostData({
-            userId: 'user-123',
-            postData,
-            postDataParameters: {} as any
-          })
-          console.log('❌ Should have failed')
-        } catch (error) {
-          const err = error as Error
-          console.log('✅ Failed as expected:', err.message)
-
-          // In the actual implementation, this error would be logged with:
-          // - errorName: 'SyntaxError'
-          // - errorMessage: the error message
-          // - triggerPostData: the original malformed JSON
-          // - triggerPostURL: the webhook URL
-
-          expect(err.name).toBe('SyntaxError')
-          expect(err.message).toContain(expectedError)
-        }
-      })
-    })
-
-    it('should handle edge cases gracefully', async () => {
-      const edgeCases = [
-        {
-          name: 'Empty post data',
-          postData: '',
-          mockError: new Error('No data to parse')
-        },
-        {
-          name: 'Null-like post data',
-          postData: 'null',
-          mockError: new Error('Invalid null data')
-        },
-        {
-          name: 'Non-object JSON',
-          postData: '"just a string"',
-          mockError: new Error('Expected object')
-        }
-      ]
-
-      edgeCases.forEach(({ name, postData, mockError }) => {
-        console.log(`\n=== Testing edge case: ${name} ===`)
-
-        mockedParseTriggerPostData.mockImplementation(() => {
-          throw mockError
-        })
-
-        try {
-          parseTriggerPostData({
-            userId: 'user-123',
-            postData,
-            postDataParameters: {} as any
-          })
-          console.log('❌ Should have failed')
-        } catch (error) {
-          console.log('✅ Handled gracefully:', (error as Error).message)
-        }
-      })
-    })
-  })
-
-  describe('Performance and Efficiency Benefits', () => {
-    it('should demonstrate network request avoidance', async () => {
-      let networkRequestCount = 0
-
-      // Track network requests
-      mockedAxios.post.mockImplementation(async () => {
-        networkRequestCount++
-        return { data: { success: true } }
-      })
-
-      console.log('\n=== Performance Test: Valid vs Invalid JSON ===')
-
-      // Test 1: Valid JSON - network request should be made
-      mockedParseTriggerPostData.mockReturnValue({ amount: 100 })
-
-      try {
-        parseTriggerPostData({
-          userId: 'user-123',
-          postData: '{"amount": <amount>}',
-          postDataParameters: {} as any
-        })
-        // In real implementation, this would proceed to make network request
-        await mockedAxios.post('https://example.com', { amount: 100 })
-        console.log('✅ Valid JSON: Network request made')
-      } catch (error) {
-        console.log('❌ Unexpected error with valid JSON')
-      }
-
-      // Test 2: Invalid JSON - no network request should be made
-      mockedParseTriggerPostData.mockImplementation(() => {
-        throw new SyntaxError('Invalid JSON')
-      })
-
-      try {
-        parseTriggerPostData({
-          userId: 'user-123',
-          postData: '{"amount": <amount>',
-          postDataParameters: {} as any
-        })
-        // Should not reach here, so no network request
-      } catch (error) {
-        console.log('✅ Invalid JSON: No network request made')
-      }
-
-      console.log(`Network requests made: ${networkRequestCount}`)
-      expect(networkRequestCount).toBe(1) // Only for valid JSON
-    })
-  })
-})
